moved progress page back to components
