+++ src/components/styles/ProgressPage.css
